+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Seller quantity condition `hasMoreSellersThan` 
+
 ## [2.1.3] - 2021-01-27
 ### Fixed
 - Avoid breaking if product context is not filled yet.