+++ CHANGELOG.md
@@ -6,9 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Added
 - Seller quantity condition `hasMoreSellersThan` 
+
+## [2.2.0] - 2021-07-28
+### Added
+- `ConditionLayoutBinding` component. 
 
 ## [2.1.3] - 2021-01-27
 ### Fixed